增加用户头像和图片比对结果缓存的功能

+++ facefoto/images.py
@@ -1,52 +1,24 @@
 import base64
 import json
 import os
-import traceback
 import time
+import traceback
+from threading import Thread
 
-import numpy
 import requests as r
 from flask import (
     Blueprint, request, jsonify, current_app
 )
 from flask_uploads import UploadSet, IMAGES, configure_uploads
 
-from facefoto.db import get_db
+from facefoto.dao import *
+from facefoto.distance_util import find_similar
+from facefoto.face import find_similar_task
 from facefoto.groups import get_group
 from facefoto.oss_util import upload_file, get_oss_url, remove_file
 
 bp = Blueprint('images', __name__)
-
-
-def get_images_by_group_id(group_id):
-    db = get_db()
-    cursor = db.cursor()
-    cursor.execute("SELECT * FROM `photo` where group_id='%s'" % group_id)
-    return cursor.fetchall()
-
-
-def get_image(id):
-    db = get_db()
-    cursor = db.cursor()
-    # 1、获得photo
-    cursor.execute('SELECT * FROM `photo` where id=%s' % id)
-    return cursor.fetchone()
-
-
-def get_image_by_oss_path(oss_path):
-    db = get_db()
-    cursor = db.cursor()
-    # 1、获得photo
-    cursor.execute("SELECT * FROM `photo` where oss_path like '%%%s'" % oss_path)
-    return cursor.fetchone()
-
-
-def delete_image_by_oss_path(oss_path):
-    db = get_db()
-    cursor = db.cursor()
-    # 1、获得photo
-    cursor.execute('DELETE FROM `photo` where oss_path=%s' % oss_path)
-    db.commit()
+task_id = 0
 
 
 # 上传图片
@@ -54,13 +26,14 @@ def delete_image_by_oss_path(oss_path):
 def images():
     """upload a photo to group."""
     global local_filename
+    global task_id
     if request.method == 'POST':
         group_id = request.form['group_id']
-        photo = request.files['photo']
-        if not photo:
-            return jsonify({'code': 400, 'msg': 'photo is required.'})
-        elif not group_id:
+        if not group_id:
             return jsonify({'code': 400, 'msg': 'group_id is required.'})
+        if 'photo' not in request.files:
+            return jsonify({'code': 400, 'msg': 'photo is required.'})
+        photo = request.files['photo']
         try:
             # 1、校验组是否存在
             group = get_group(group_id)
@@ -89,15 +62,11 @@ def images():
                 # 4、上传图片到oss
                 oss_url = upload_file(oss_filename, local_filename)
                 # 5、保存photo数据到db
-
-                db = get_db()
-                cursor = db.cursor()
-                cursor.execute(
-                    "insert into `photo` (name, oss_path, group_id, feature) VALUES ('%s','%s','%d','%s')" % (
-                        filename, oss_filename, int(group_id), json.dumps(body)))
-                image_id = cursor.lastrowid
-                db.commit()
-
+                image_id = add_image(filename, oss_filename, int(group_id), json.dumps(body))
+                task_id += 1
+                Thread(target=find_similar_task,
+                       args=(current_app._get_current_object(), "image_%s" % task_id, group_id, image_id,)
+                       ).start()
                 return jsonify(
                     {'code': 200, 'image_id': image_id, 'url': oss_url,
                      'url_express': current_app.config['OSS_URL_EXPIRES'],
@@ -109,7 +78,13 @@ def images():
             return jsonify({'code': 500, 'error': '{0}'.format(e)})
         finally:
             # 6、删除临时图片
-            os.remove(local_filename)
+            try:
+                if os.path.isfile(local_filename):
+                    os.remove(local_filename)
+            except FileNotFoundError:
+                print("delete not exits file")
+            except Exception:
+                traceback.print_exc()
     else:
         return jsonify({'code': 400, 'msg': 'upload image failed.'})
 
@@ -153,14 +128,8 @@ def images_from_oss():
             if 'body' in response:
                 body = response['body']
                 # 5、保存photo数据到db
-                db = get_db()
-                cursor = db.cursor()
-                cursor.execute(
-                    "insert into `photo` (name, oss_path, group_id, feature) VALUES ('%s','%s','%s','%s')" % (
-                        filename, oss_filename.split(current_app.config['BUCKET_NAME'])[-1], group_id,
-                        json.dumps(body)))
-                image_id = cursor.lastrowid
-                db.commit()
+                image_id = add_image(filename, oss_filename.split(current_app.config['BUCKET_NAME'])[-1], group_id,
+                                     json.dumps(body))
 
                 return jsonify(
                     {'code': 200, 'image_id': image_id, 'url': oss_url,
@@ -191,11 +160,11 @@ def search_by_photo():
     global local_filename
     if request.method == 'POST':
         group_id = request.form['group_id']
-        photo = request.files['photo']
-        if not photo:
-            return jsonify({'code': 400, 'msg': 'photo is required.'})
-        elif not group_id:
+        if not group_id:
             return jsonify({'code': 400, 'msg': 'group_id is required.'})
+        if 'photo' not in request.files:
+            return jsonify({'code': 400, 'msg': 'photo is required.'})
+        photo = request.files['photo']
         try:
             # 1、校验组是否存在
             group = get_group(group_id)
@@ -226,29 +195,17 @@ def search_by_photo():
 
             if 'body' in response:
                 body = response['body']
-                for b in body:
-                    features = b['features']
-                    matched_images = {}
-                    for one in target_images:
-                        print("target image :{0}".format(one[1]))
-                        one_body = json.loads(one[4])
-                        for one_b in one_body:
-                            one_features = one_b['features']
-                            dist = get_euclidean_distance(features, one_features)
-                            if dist < limit and filename != one[1]:
-                                print("image match :{0}".format(one[1]))
-                                matched_images[one] = dist
-                                break
-                    sorted_matched_images = sort_by_value(matched_images)
-                    # print(sorted_matched_images)
-                    result_body = []
-                    for matched in sorted_matched_images:
-                        oss_path = matched[2]
-                        expires = current_app.config['OSS_URL_EXPIRES']
-                        oss_url = get_oss_url(oss_path, expires)
-                        result_body.append({'id': matched[0], 'name': matched[1], 'url': oss_url, 'expires': expires})
+                image = (None, filename, None, int(group_id), json.dumps(body))
+                # 6、比对feature值
+                sorted_matched_images = find_similar(image, target_images)
+                result_body = []
+                for matched in sorted_matched_images:
+                    oss_path = matched[2]
+                    expires = current_app.config['OSS_URL_EXPIRES']
+                    oss_url = get_oss_url(oss_path, expires)
+                    result_body.append({'id': matched[0], 'name': matched[1], 'url': oss_url, 'expires': expires})
 
-                    return jsonify({'code': 200, 'data': result_body})
+                return jsonify({'code': 200, 'data': result_body})
             else:
                 return jsonify(response)
         except Exception as e:
@@ -277,42 +234,21 @@ def search_by_oss_path():
             oss_path = data['oss_path']
             print(oss_path)
 
-            limit = float(current_app.config['FEATURES_MAX_MATCH_LIMIT'])
-            limit = 2 * (1 - limit)
-            print("limit {0}".format(limit))
             # 1、获得photo
             image = get_image_by_oss_path(oss_path)
             # 2、获得指定group_id下的所有photo
             target_images = get_images_by_group_id(group_id)
-            # 3、比对features值
-            matched_images = {}
             if not image:
                 return jsonify({'code': 400, 'msg': 'cant not find image by oss_path{0}'.format(oss_path)})
             if not target_images:
                 return jsonify({'code': 400, 'msg': 'cant not find any image by group id {0}'.format(group_id)})
 
-            body = json.loads(image[4])
-            print("src image :{0}".format(image[1]))
-            for b in body:
-                features = b['features']
-
-                for one in target_images:
-                    print("target image :{0}".format(one[1]))
-                    one_body = json.loads(one[4])
-                    for one_b in one_body:
-                        one_features = one_b['features']
-                        dist = get_euclidean_distance(features, one_features)
-                        if dist < limit and image[1] != one[1]:
-                            print("image match :{0}".format(one[1]))
-                            matched_images[one] = dist
-                            break
-                sorted_matched_images = sort_by_value(matched_images)
-                # print(sorted_matched_images)
-                result_body = []
-                for matched in sorted_matched_images:
-                    result_body.append({'id': matched[0], 'name': matched[1]})
-
-                return jsonify({'code': 200, 'data': result_body})
+            # 3、比对features值
+            sorted_matched_images = find_similar(image, target_images)
+            result_body = []
+            for matched in sorted_matched_images:
+                result_body.append({'id': matched[0], 'name': matched[1]})
+            return jsonify({'code': 200, 'data': result_body})
         except Exception as e:
             return jsonify({'code': 500, 'error': "{0}".format(e)})
 
@@ -327,54 +263,25 @@ def search():
             return jsonify({'code': 400, 'msg': 'image_id is required.'})
         elif not group_id:
             return jsonify({'code': 400, 'msg': 'group_id is required.'})
-        limit = float(current_app.config['FEATURES_MAX_MATCH_LIMIT'])
-        limit = 2 * (1 - limit)
-        print("limit {0}".format(limit))
         # 1、获得photo
         image = get_image(image_id)
         # 2、获得指定group_id下的所有photo
         target_images = get_images_by_group_id(group_id)
-        # 3、比对features值
-        matched_images = {}
         if not image:
             return jsonify({'code': 400, 'msg': 'cant not find image by id {0}'.format(image_id)})
-
-        body = json.loads(image[4])
-        print("src image :{0}".format(image[1]))
-        for b in body:
-            features = b['features']
-            if not target_images:
-                return jsonify({'code': 400, 'msg': 'cant not find any image by group id {0}'.format(group_id)})
-
-            for one in target_images:
-                print("target image :{0}".format(one[1]))
-                one_body = json.loads(one[4])
-                for one_b in one_body:
-                    one_features = one_b['features']
-                    dist = get_euclidean_distance(features, one_features)
-                    if dist < limit and image[1] != one[1]:
-                        print("image match :{0}".format(one[1]))
-                        matched_images[one] = dist
-                        break
-            sorted_matched_images = sort_by_value(matched_images)
-            # print(sorted_matched_images)
-            result_body = []
-            for matched in sorted_matched_images:
-                oss_path = matched[2]
-                expires = current_app.config['OSS_URL_EXPIRES']
-                oss_url = get_oss_url(oss_path, expires)
-                result_body.append({'id': matched[0], 'name': matched[1], 'url': oss_url, 'expires': expires})
-
-            return jsonify({'code': 200, 'data': result_body})
+        if not target_images:
+            return jsonify({'code': 400, 'msg': 'cant not find any image by group id {0}'.format(group_id)})
+        # 3、比对features值
+        sorted_matched_images = find_similar(image, target_images)
+        result_body = []
+        for matched in sorted_matched_images:
+            oss_path = matched[2]
+            expires = current_app.config['OSS_URL_EXPIRES']
+            oss_url = get_oss_url(oss_path, expires)
+            result_body.append({'id': matched[0], 'name': matched[1], 'url': oss_url, 'expires': expires})
+        return jsonify({'code': 200, 'data': result_body})
     except Exception as e:
         return jsonify({'code': 500, 'error': "{0}".format(e)})
-
-
-def sort_by_value(d):
-    items = d.items()
-    backitems = [[v[1], v[0]] for v in items]
-    backitems.sort(reverse=True)
-    return [backitems[i][1] for i in range(0, len(backitems))]
 
 
 # 根据id获得图片url
@@ -425,20 +332,8 @@ def delete(image_id):
 
         remove_file(oss_filename)
 
-        db = get_db()
-        cursor = db.cursor()
-        cursor.execute("DELETE FROM `photo` where id ='%d'" % image_id)
-        db.commit()
+        delete_image(image_id)
         return jsonify(
             {'code': 200, 'msg': 'image [{0}] delete success'.format(image[1])})
     except Exception as e:
         return jsonify({'code': 500, 'error': "{0}".format(e)})
-
-
-# 计算两个特征值的欧式距离，此例中欧式距离最大值为2，距离越小，人像越相似
-def get_euclidean_distance(src, target):
-    v1 = numpy.array(src)
-    v2 = numpy.array(target)
-    dist = numpy.sqrt(numpy.sum(numpy.square(v1 - v2)))
-    print(dist)
-    return dist
